+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -169,7 +169,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-11T12:55:42.903Z",
+    "2026-06-12T06:27:06.026Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {